chore(php) Update to PHP 7.
