+++ Bin/Generate.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 /**
  * Hoa
@@ -45,7 +45,6 @@ use Hoa\Test;
  * Automatically generate test suites based on examples in API documentations
  * or from contracts written in Praspel.
  *
- * @copyright  Copyright © 2007-2017 Hoa community
  * @license    New BSD License
  */
 class Generate extends Console\Dispatcher\Kit
@@ -70,10 +69,10 @@ class Generate extends Console\Dispatcher\Kit
      *
      * @return  int
      */
-    public function main()
+    public function main(): int
     {
-        $directoryToScan       = null;
-        $namespaceToScan       = null;
+        $directoryToScan = null;
+        $namespaceToScan = null;
 
         while (false !== $c = $this->getOption($v)) {
             switch ($c) {
@@ -111,7 +110,7 @@ class Generate extends Console\Dispatcher\Kit
             return $this->usage();
         }
 
-        return;
+        return 0;
     }
 
     /**
@@ -119,7 +118,7 @@ class Generate extends Console\Dispatcher\Kit
      *
      * @return  int
      */
-    public function usage()
+    public function usage(): int
     {
         echo
             'Usage   : test:generate <options>', "\n",
@@ -131,7 +130,7 @@ class Generate extends Console\Dispatcher\Kit
                 'help' => 'This help.'
             ]), "\n";
 
-        return;
+        return 0;
     }
 }
