+++ Bin/Run.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 /**
  * Hoa
@@ -39,16 +39,15 @@ namespace Hoa\Test\Bin;
 use Hoa\Consistency;
 use Hoa\Console;
 use Hoa\File;
-use Kitab;
 use Hoa\Protocol;
 use Hoa\Test;
+use Kitab;
 
 /**
  * Class Hoa\Test\Bin\Run.
  *
  * Run tests.
  *
- * @copyright  Copyright © 2007-2017 Hoa community
  * @license    New BSD License
  */
 class Run extends Console\Dispatcher\Kit
@@ -80,7 +79,7 @@ class Run extends Console\Dispatcher\Kit
      *
      * @return  int
      */
-    public function main()
+    public function main(): int
     {
         $directories         = [];
         $files               = [];
@@ -380,21 +379,21 @@ class Run extends Console\Dispatcher\Kit
         $processus->on('input', function ($bucket) {
             return false;
         });
-        $processus->on('output', function ($bucket) {
+        $processus->on('output', function ($bucket): void {
             $data = $bucket->getData();
 
             echo $data['line'], "\n";
 
             return;
         });
-        $processus->on('stop', function ($bucket) {
+        $processus->on('stop', function ($bucket): void {
             // Wait atoum to finish its sub-children.
             sleep(1);
             exit($bucket->getSource()->getExitCode());
         });
         $processus->run();
 
-        return;
+        return 0;
     }
 
     /**
@@ -402,7 +401,7 @@ class Run extends Console\Dispatcher\Kit
      *
      * @return  int
      */
-    public function usage()
+    public function usage(): int
     {
         echo
             'Usage   : test:run <options>', "\n",
@@ -428,7 +427,7 @@ class Run extends Console\Dispatcher\Kit
             '    * namespace,', "\n",
             '    * tags.', "\n";
 
-        return;
+        return 0;
     }
 }
