--- Report/Cli/Fields/Uncompleted.php
@@ -1,81 +1,0 @@
-<?php
-
-/**
- * Hoa
- *
- *
- * @license
- *
- * New BSD License
- *
- * Copyright © 2007-2017, Hoa community. All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *     * Redistributions of source code must retain the above copyright
- *       notice, this list of conditions and the following disclaimer.
- *     * Redistributions in binary form must reproduce the above copyright
- *       notice, this list of conditions and the following disclaimer in the
- *       documentation and/or other materials provided with the distribution.
- *     * Neither the name of the Hoa nor the names of its contributors may be
- *       used to endorse or promote products derived from this software without
- *       specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDERS AND CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- */
-
-namespace Hoa\Test\Report\Cli\Fields;
-
-use atoum\report\fields;
-
-class Uncompleted extends fields\runner\tests\uncompleted\cli
-{
-    public function __toString()
-    {
-        $string = '';
-
-        if (null !== $this->runner) {
-            $uncompletedMethods      = $this->runner->getScore()->getUncompletedMethods();
-            $sizeOfUncompletedMethod = sizeof($uncompletedMethods);
-
-            if (0 < $sizeOfUncompletedMethod) {
-                $string .=
-                    $this->titlePrompt .
-                    sprintf(
-                        $this->locale->_('%s:'),
-                        $this->titleColorizer->colorize(sprintf($this->locale->__('There is %d uncompleted test case', 'There are %d uncompleted test cases', $sizeOfUncompletedMethod), $sizeOfUncompletedMethod))
-                    ) . "\n";
-
-                foreach ($uncompletedMethods as $uncompletedMethod) {
-                    $string .=
-                        $this->methodPrompt .
-                        sprintf(
-                            $this->locale->_('%s:'),
-                            $this->methodColorizer->colorize(sprintf('%s::%s() with exit code %d', $uncompletedMethod['class'], $uncompletedMethod['method'], $uncompletedMethod['exitCode']))
-                        ) . "\n";
-
-                    $lines   = explode(PHP_EOL, trim($uncompletedMethod['output']));
-                    $string .= $this->outputPrompt . 'output(' . strlen($uncompletedMethod['output']) . ') "' . array_shift($lines);
-
-                    foreach ($lines as $line) {
-                        $string .= "\n" . $this->outputPrompt . $line;
-                    }
-
-                    $string .= '"' . "\n";
-                }
-            }
-        }
-
-        return $string;
-    }
-}
