+++ Source/Report/Cli/Fields/Uncompleted.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * Hoa
  *
@@ -33,28 +35,48 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-
 namespace Hoa\Test\Report\Cli\Fields;
 
 use atoum\report\fields;
 
-class Duration extends fields\test\duration\cli
+class Uncompleted extends fields\runner\tests\uncompleted\cli
 {
     public function __toString()
     {
-        return
-            $this->prompt .
-            sprintf(
-                $this->locale->_('%1$s: %2$s.'),
-                $this->titleColorizer->colorize($this->locale->_('Duration')),
-                $this->durationColorizer->colorize(
-                    null === $this->value
-                        ? $this->locale->_('unknown')
-                        : sprintf(
-                            $this->locale->__('%4.6f second', '%4.6f seconds', $this->value),
-                            $this->value
-                        )
-                )
-            ) . "\n";
+        $string = '';
+
+        if (null !== $this->runner) {
+            $uncompletedMethods      = $this->runner->getScore()->getUncompletedMethods();
+            $sizeOfUncompletedMethod = sizeof($uncompletedMethods);
+
+            if (0 < $sizeOfUncompletedMethod) {
+                $string .=
+                    $this->titlePrompt .
+                    sprintf(
+                        $this->locale->_('%s:'),
+                        $this->titleColorizer->colorize(sprintf($this->locale->__('There is %d uncompleted test case', 'There are %d uncompleted test cases', $sizeOfUncompletedMethod), $sizeOfUncompletedMethod))
+                    ) . "\n";
+
+                foreach ($uncompletedMethods as $uncompletedMethod) {
+                    $string .=
+                        $this->methodPrompt .
+                        sprintf(
+                            $this->locale->_('%s:'),
+                            $this->methodColorizer->colorize(sprintf('%s::%s() with exit code %d', $uncompletedMethod['class'], $uncompletedMethod['method'], $uncompletedMethod['exitCode']))
+                        ) . "\n";
+
+                    $lines   = explode(PHP_EOL, trim($uncompletedMethod['output']));
+                    $string .= $this->outputPrompt . 'output(' . strlen($uncompletedMethod['output']) . ') "' . array_shift($lines);
+
+                    foreach ($lines as $line) {
+                        $string .= "\n" . $this->outputPrompt . $line;
+                    }
+
+                    $string .= '"' . "\n";
+                }
+            }
+        }
+
+        return $string;
     }
 }
